A good start
Updating the pawn to work with the new Scout ship.

The fans spin more quickly while you're moving, and return to idle speed when you stop.

Adjusted the GunOffset to accomodate the new ship size.

I have a solution to make the gun attachments face the firing direction, but it doesn't look great (because the guns weren't modeled with turret ability in mind).  Might just leave them in place, forcing the player to only use those guns in the forward direction.