Progress on fixing valid pose bug
Basically having an issue attaching the fans/guns to the ship body, because the body itself is simulating physics while the attachments are not.

I figure the solution is physics constraints, but as you can see in the CommonLibrary.h file I added with this commit, the constraint settings have been deprecated without documentation.

I'll need to find the replacement settings to get that working.