Added Combat Text Component
Converted Combat Text Component code from Ethereal Legends.  The status effect feature as seen in Ethereal is not yet functional and may not end up being supported in ERHD, but it successfully displays damage and text in worldspace above the parent actor, which I figured was more important at the moment.