Upgrade sequencing
The unlock and upgrade sequence for Lasers, Magnet, and Bombs is now fully functional.  Health upgrade sequence was already working.  Added features to the pickups so that they'll display info and destroy themselves if you attempt a pickup when the feature is still locked or if you have the max capacity already.  Added updated description text to the hangar menu inserts that display different information at different times during the upgrade sequence.