Started work on the first enemy
including "Enemy Manager", "Enemy Master", and "Drone" as the first enemy.

procedural spawn functionality works, and the classes are parented more or less like how I did it for Ethereal and provide some of the same functionality.  Should be a good framework to build out the rest of the enemies.